refactor : move `Notice` type to shared layer
