--- src/features/notice/types/notice.ts
@@ -1,15 +1,0 @@
-export type NoticeType = 'UPCOMING' | 'NOTIFY';
-
-export type NoticeListType = NoticeType | 'ONGOING' | 'COMPLETED';
-
-export type Notice = {
-  id: number;
-  title: string;
-  startAt: string;
-  endAt: string;
-  noticeType: NoticeType;
-};
-
-export type NoticeList = Omit<Notice, 'noticeType'> & {
-  noticeStatus: NoticeListType;
-};
