+++ src/shared/types/index.ts
@@ -4,3 +4,5 @@ export * from './election';
 export * from './path';
 export * from './user';
 export * from './vote';
+export * from './util';
+export * from './notice';
